% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnplasso_lm.R
\name{bnplasso.lm}
\alias{bnplasso.lm}
\title{Fit linear regression models with a nonparametric Bayesian Lasso prior}
\usage{
bnplasso.lm(
  X,
  y,
  intercept = TRUE,
  prior = NULL,
  a = NULL,
  b = NULL,
  alpha = NULL,
  variance.prior.type = NULL,
  max.iters = 6000L,
  burn.in = 1000L,
  thin = 1L,
  float = FALSE
)
}
\arguments{
\item{X}{A matrix of predictors of dimension \eqn{n}-by-\eqn{p}, where each 
of the \eqn{n} rows is an observation vector.}

\item{y}{Response variable. It should be a numeric vector size \eqn{n}.}

\item{intercept}{Logical. If \code{TRUE}, an intercept term is included in 
the model; otherwise, the intercept is integrated out. If \code{TRUE}, the 
prior on the intercept would be a non-informative prior of the form 
\eqn{p(\mu)\propto 1}. Default is \code{TRUE}.}

\item{prior}{A character string denoting which type of shrinkage prior
should be employed. The options are either: (a) \code{"bnp.lasso"} which
implements the nonparametric Bayesian Lasso as in Marin et al. (2025+), 
(b) \code{"b.lasso"} which implements the Bayesian Lasso as in Park and 
Casella (2008), or (c) \code{"b.adapt.lasso"} which implements the Bayesian
adaptive Lasso as in Leng et al. (2014). Default is \code{"bnp.lasso"}.}

\item{a}{A positive scalar. If \code{prior = "bnp.lasso"}, it 
corresponds to the \bold{shape} parameter in the gamma distribution used 
as a centering measure in the DP prior. If \code{prior = "b.lasso"}, 
it corresponds to the \bold{shape} parameter in the gamma distribution used 
as prior in the shrinkage parameter, \eqn{\lambda^{2}}. If 
\code{prior = "b.adapt.lasso"}, it corresponds to the \bold{shape}
parameter in the gamma distribution used as prior in the shrinkage 
parameters, \eqn{\lambda_{j}^{2}}, for \eqn{j\in\{1,\dots,p\}}. If not 
provided, the function will attempt to determine an appropriate value.}

\item{b}{A positive scalar. If \code{prior = "bnp.lasso"}, it 
corresponds to the \bold{rate} parameter in the gamma distribution used 
as a centering measure in the DP prior. If \code{prior = "b.lasso"}, 
it corresponds to the \bold{rate} parameter in the gamma distribution used 
as prior in the shrinkage parameter, \eqn{\lambda^{2}}. If 
\code{prior = "b.adapt.lasso"}, it corresponds to the \bold{rate}
parameter in the gamma distribution used as prior in the shrinkage 
parameters, \eqn{\lambda_{j}^{2}}, for \eqn{j\in\{1,\dots,p\}}. If not 
provided, the function will attempt to determine an appropriate value.}

\item{alpha}{A positive scalar. If \code{prior = "bnp.lasso"}, it 
corresponds to the \bold{concentration} parameter in the DP prior. If 
\code{prior = "b.lasso"} or \code{prior = "b.adapt.lasso"},
the argument is ignored. If not provided, the function will attempt to 
determine an appropriate value.}

\item{variance.prior.type}{A character string denoting whether the variance 
on the sampling variance should be an independent-type prior or a 
conjugate-type prior. See Moran et al. (2019) for details. The options are 
either \code{"independent"} or \code{"conjugate"}. If not provided, the 
function will attempt to determine an appropriate value.}

\item{max.iters}{A positive integer corresponding to the total number of 
MCMC iterations. Default is 6000.}

\item{burn.in}{A positive integer corresponding to the number of draws 
discarded as burn-in. It should be smaller than \code{max.iters}.
Default is 1000.}

\item{thin}{A positive integer specifying the period for saving samples.
Default is 1.}

\item{float}{Logical. If \code{TRUE}, some internal routines will use single 
point precision for improved computational efficiency at the expense of 
numerical accuracy. If \code{FALSE}, double point precision will be used. 
Default is \code{FALSE}.}
}
\value{
An object of S3 class, "lmBayes", containing:
\itemize{
  \item \code{Post.beta}: A matrix of size \code{n.draws}-by-\code{n.preds}, 
  where each row is a posterior draw of the regression coefficients.
  \item \code{Post.sigma2}: A vector of size \code{n.draws}, where each 
  element is a posterior draw of the sampling variance. 
  \item \code{Post.tau2}: A matrix of size \code{n.draws}-by-\code{n.preds}, 
  where each row is a posterior draw of the latent parameters 
  \eqn{\tau_{j}^{2}}.
  \item \code{Post.lambda2}: A matrix of size 
  \code{n.draws}-by-\code{n.preds}, where each row is a posterior draw of the
  shrinkage parameters \eqn{\lambda_{j}^{2}}.
  \item \code{Post.clust_idx}: If \code{prior = "bnp.lasso"}, a matrix of 
  size \code{n.draws}-by-\code{n.preds}, where each row indicates to which 
  cluster the regression coefficients belong to.
  \item \code{Post.K}: If \code{prior = "bnp.lasso"}, a vector of size 
  \code{n.draws}, where each element indicates the number of clusters in the 
  corresponding MCMC iteration.
  \item \code{Post.mu}: If \code{intercept = "TRUE"}, a vector of size 
  \code{n.draws}, where each entry is a posterior draw of the intercept term.
  \item \code{elapsed}: The elapsed (wall-clock) time of the MCMC sampler.
  \item \code{a}: If \code{prior = "bnp.lasso"}, the \bold{shape} parameter 
  in the gamma distribution used as a centering measure in the DP prior. 
  If \code{prior = "b.lasso"}, the \bold{shape} parameter in the gamma 
  distribution used as prior in the shrinkage parameter, \eqn{\lambda^{2}}. 
  If \code{prior = "b.adapt.lasso"}, the \bold{shape} parameter in the gamma 
  distribution used as prior in the shrinkage parameters, 
  \eqn{\lambda_{j}^{2}}, for \eqn{j\in\{1,\dots,p\}}.
  \item \code{b}: If \code{prior = "bnp.lasso"}, the \bold{rate} parameter 
  in the gamma distribution used as a centering measure in the DP prior. 
  If \code{prior = "b.lasso"}, the \bold{rate} parameter in the gamma 
  distribution used as prior in the shrinkage parameter, \eqn{\lambda^{2}}. 
  If \code{prior = "b.adapt.lasso"}, the \bold{rate} parameter in the gamma 
  distribution used as prior in the shrinkage parameters, 
  \eqn{\lambda_{j}^{2}}, for \eqn{j\in\{1,\dots,p\}}.
  \item \code{alpha}: If \code{prior = "bnp.lasso"}, the \bold{concentration} 
  parameter in the Dirichlet process prior.
  \item \code{intercept}: Whether an intercept was included in the model.
  \item \code{variance.prior.type}: Whether the variance on the sampling 
  variance was an independent-type prior or a conjugate-type prior.
  \item \code{max.iters}: The total number of MCMC iterations.
  \item \code{burn.in}: The number of draws discarded as burn-in.
  \item \code{thin}: The period for saving draws.
  \item \code{n.obs}: The sample size.   
  \item \code{n.preds}: The number of predictors.
  \item \code{n.draws}: The number of posterior draws after burn-in and
  thinning.
  \item \code{X}: Matrix of predictors.
  \item \code{y}: Vector of responses.
  \item \code{loglik}: Matrix of size \code{n.draws}-by-\code{n.obs} with the 
  log-likelihood of each observation at each MCMC iteration.
  \item \code{post.pred.fitted.values}: A matrix of size 
  \code{n.draws}-by-\code{n.obs}, where each row is a draw from the posterior 
  predictive distribution of the fitted values.
  \item \code{post.pred.residuals}: A matrix of size 
  \code{n.draws}-by-\code{n.obs}, where each row is a draw from the posterior 
  predictive distribution of the residuals.  
}
}
\description{
This function fits linear regression models with a nonparametric 
Bayesian Lasso prior as in Marin et al. (2025+).
}
\references{
C. Leng, MN. Tran, and D. Nott (2014), Bayesian adaptive Lasso. 
\emph{Ann Inst Stat Math}, 66:221-244

S. Marin, B. Long,and A. H. Westveld (2025+), Adaptive Shrinkage with a 
Nonparametric Bayesian Lasso.\emph{Journal of Computational and Graphical 
Statistics}. doi:10.1080/10618600.2025.2572327

G. E. Moran, V. Rockova, and E. I. George (2019), Variance Prior Forms for 
High-Dimensional Bayesian Variable Selection. \emph{Bayesian Analysis},
14(4):1091-1119.

T. Park, and G. Casella (2008), The Bayesian Lasso. 
\emph{Journal of the American Statistical Association}, 103(482):681-686.
}
\author{
Santiago Marin
}
