% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnplasso-package.R
\docType{package}
\name{bnplasso}
\alias{bnplasso-package}
\alias{bnplasso}
\title{bnplasso: Fit linear regression models with the Nonparametric Bayesian Lasso}
\description{
Fit linear regression models with the Nonparametric Bayesian Lasso, a novel 
and adaptive shrinkage prior for Bayesian regression and variable selection. 
Posterior inference is conducted via Markov chain Monte Carlo (MCMC) as 
described in Marin et al. (2025+) 
<https://doi.org/10.48550/arXiv.2311.03644>.
}
\references{
S. Marin, B. Long,and A. H. Westveld (2025+), Adaptive Shrinkage with a 
Nonparametric Bayesian Lasso. \emph{arXiv}, {2411.08262}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marinsantiago/bnplasso/issues}
  \item Report bugs at \url{https://github.com/marinsantiago/bnplasso/issues}
}

}
\author{
\strong{Maintainer}: Santiago Marin \email{santiago.marinardila@anu.edu.au} (\href{https://orcid.org/0009-0006-0086-3721}{ORCID})

}
