% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclust.R
\name{coclust.point}
\alias{coclust.point}
\title{Posterior co-clustering plot}
\usage{
coclust.point(
  S,
  return.mat = FALSE,
  viridis.pal = "C",
  main = NULL,
  axis.idx = NULL,
  xy.labs = NULL
)
}
\arguments{
\item{S}{A vector of size \code{n.preds}, indicating to which cluster the 
corresponding regression coefficients belong to.}

\item{return.mat}{logical. Whether the matrix of posterior co-clustering 
probabilities should be returned. Default is \code{FALSE}}

\item{viridis.pal}{Character describing which "viridis" palette should be
employed. Options are "A", "B", "C", "D", "E", "F", "G", and "H".}

\item{main}{Title of the plot.}

\item{axis.idx}{Sequence of number to appear in the axis of the plot.}

\item{xy.labs}{A title for the x and y axes.}
}
\description{
Plots the posterior co-clustering of the regression coefficients
}
\author{
Santiago Marin
}
