% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coclust.R
\name{coclust.probs}
\alias{coclust.probs}
\title{Posterior co-clustering probabilities plot}
\usage{
coclust.probs(
  S.mcmc,
  return.mat = FALSE,
  viridis.pal = "C",
  main = NULL,
  axis.idx = NULL,
  xy.labs = NULL
)
}
\arguments{
\item{S.mcmc}{A matrix of size \code{n.draws}-by-\code{n.preds}, where each 
row indicates to which cluster the regression coefficients belong to.}

\item{return.mat}{logical. Whether the matrix of posterior co-clustering 
probabilities should be returned. Default is \code{FALSE}}

\item{viridis.pal}{Character describing which "viridis" palette should be
employed. Options are "A", "B", "C", "D", "E", "F", "G", and "H".}

\item{main}{Title of the plot.}

\item{axis.idx}{Sequence of number to appear in the axis of the plot.}

\item{xy.labs}{A title for the x and y axes.}
}
\description{
This function plots the posterior co-clustering probabilities of the 
regression coefficients
}
\author{
Santiago Marin
}
