% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{coef.spmBayes}
\alias{coef.spmBayes}
\alias{coefficients.spmBayes}
\title{Mean parameters}
\usage{
\method{coef}{spmBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{'spmBayes'}.}

\item{...}{Further arguments passed to.}
}
\value{
A data.frame with summary statistics of the posterior distribution of
the mean parameters.
}
\description{
Extracts the posterior distribution of the mean parameters from an
object of class \code{spmBayes}.
}
\author{
Santiago Marin
}
