% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elppd.R
\name{elppd}
\alias{elppd}
\title{Expected log pointwise predictive density (elppd)}
\usage{
elppd(object, X.new, y.new)
}
\arguments{
\item{object}{An object of class \code{lmBayes}.}

\item{X.new}{A matrix of predictors from the test data, where each row is an 
observation.}

\item{y.new}{A vector of responses from the test data of size 
\eqn{n_{\text{test}}}.}
}
\value{
The elppd as a numeric scalar.
}
\description{
This function computes the expected log pointwise predictive density (elppd) 
for an object of class \code{lmBayes}. Given a test (held-out) set, 
\eqn{\{y_{\text{test},i},\,
\mathbf{x}_{\text{test},i}\}_{i=1}^{n_{\text{test}}}}, the elppd is defined 
as 
\deqn{\text{elppd} = \frac{1}{n_{\text{test}}}\sum_{i=1}^{n_{\text{test}}}
\log\left(\frac{1}{S}\sum_{s=1}^{S}p\left({y}_{\text{test}, i} | 
\mathbf{x}_{\text{test}, i}^{_{'}}\boldsymbol{\beta}_{(s)}, 
\sigma^{2}_{(s)}\right)\right),}
where \eqn{p()} is the likelihood of the observed data and 
\eqn{\boldsymbol{\beta}_{(s)}} and \eqn{\sigma^{2}_{(s)}} are the 
\eqn{s}-th draws of the coefficient vector and the sampling variance in the 
MCMC algorithm, respectively.
}
\references{
A. Gelman, J. Carlin, H. Stern, D. Dunson, and A. Vehtari (2013), 
\emph{Bayesian Data Analysis}, 3. Chapman & Hall/CRC.
}
\author{
Santiago Marin
}
