% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{fitted.spmBayes}
\alias{fitted.spmBayes}
\alias{fitted.values.spmBayes}
\title{Fitted values}
\usage{
\method{fitted}{spmBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{'spmBayes'}.}

\item{...}{Further arguments passed to.}
}
\value{
A data.frame with summary statistics of the posterior predictive 
fitted values.
}
\description{
Extracts the posterior predictive fitted values from an object of class
\code{spmBayes}.
}
\author{
Santiago Marin
}
