% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{get.partition}
\alias{get.partition}
\title{Recover a partition}
\usage{
get.partition(S.mcmc, loss = "VI")
}
\arguments{
\item{S.mcmc}{A matrix of size \code{n.draws}-by-\code{n.preds}, where each 
row indicates to which cluster the regression coefficients belong to.}

\item{loss}{A loss function defined on the space of partitions. It can be 
either the variation of information loss function (\code{"VI"}) or the
Binder loss function (\code{"Binder"}). Default is \code{"VI"}. See 
Wade and Ghahramani (2018) for additional details.}
}
\description{
This function recovers the partition of the regression coefficients induced 
by the nonparametric Bayesian Lasso. This function is based on the 
implementation from the \code{"BNPmix"} package.
}
\references{
S. Wade and Z. Ghahramani (2018), Bayesian cluster analysis: Point 
estimation and credible balls (with discussion). \emph{Bayesian Analysis},
13(2):559-626.
}
\author{
Santiago Marin
}
