% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{plot.lmBayes}
\alias{plot.lmBayes}
\title{Plot the results and diagnostics for an object of class \code{lmBayes}}
\usage{
\method{plot}{lmBayes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{'lmBayes'}.}

\item{...}{Further arguments passed to.}
}
\description{
Produces posterior predictive and residual diagnostic plots for an object of 
class \code{'lmBayes'}.
}
\author{
Santiago Marin
}
