% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{plot.spmBayes}
\alias{plot.spmBayes}
\title{Plot the results and diagnostics for an object of class \code{spmBayes}}
\usage{
\method{plot}{spmBayes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{'spmBayes'}.}

\item{...}{Further arguments passed to.}
}
\description{
Produces posterior predictive and residual diagnostic plots for an object of 
class \code{'spmBayes'}.
}
\author{
Santiago Marin
}
