% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_estimates.R
\name{point.estimates}
\alias{point.estimates}
\title{Point estimates of the regression coefficients for an object of class 
\code{lmBayes}}
\usage{
point.estimates(object, type = "sn", retain = "mode")
}
\arguments{
\item{object}{An object of class 'lmBayes'.}

\item{type}{A character string denoting which algorithm should be used to
recover the point estimates. The options are: (i) \code{"sn"} for the
scaled neighborhood criterion from Li and Lin (2010), (ii) 
\code{"cred.int"} which excludes a predictor if the 50\% credible interval 
contains zero, (iii) \code{"post.mode"} for the posterior mode, or (iv)
\code{"post.mean"} for the posterior mean. Default is \code{"sn"}.}

\item{retain}{A character string denoting which posterior summary should be 
retained if \code{type = "sn"} or \code{type = "cred.int"}. The options 
are: (i) \code{"mode"} to retain the posterior mode or (ii) \code{"mean"} 
to retain the posterior mean.}
}
\value{
A numeric vector containing the point estimates of the regression
coefficients.
}
\description{
Compute point estimates of the regression coefficients for an object of class 
\code{lmBayes}.
}
\author{
Santiago Marin
}
