% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{predict.lmBayes}
\alias{predict.lmBayes}
\title{Posterior predictive distribution for new data}
\usage{
\method{predict}{lmBayes}(object, X.new, ...)
}
\arguments{
\item{object}{An object of class \code{'lmBayes'}.}

\item{X.new}{A new matrix of predictors, where each row is a new observation.}

\item{...}{Further arguments passed to.}
}
\value{
A matrix where each row corresponds to an MCMC draw and each column 
to an observation in the new data.
}
\description{
Compute the posterior predictive distribution for an object of class 
\code{lmBayes}.
}
\author{
Santiago Marin
}
