% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{print.spmBayes}
\alias{print.spmBayes}
\title{Print the results for an object of class \code{spmBayes}}
\usage{
\method{print}{spmBayes}(x, ...)
}
\arguments{
\item{x}{An object of class \code{spmBayes}.}

\item{...}{Further arguments passed to.}
}
\description{
Print the results for an object of class \code{spmBayes}
}
\author{
Santiago Marin
}
