% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{psis.loo}
\alias{psis.loo}
\title{PSIS-LOO}
\usage{
psis.loo(object, comp.r_eff = TRUE)
}
\arguments{
\item{object}{An object of class \code{'lmBayes'}.}

\item{comp.r_eff}{Logical. Whether the function should estimate the relative 
effective sample size for the likelihood of each observation. If 
\code{FALSE}, all the relative effective sample sizes are set to one.
Default is \code{TRUE}.}
}
\value{
The estimated PSIS-LOO.
}
\description{
This function computes the Pareto-smoothed importance sampling leave-one-out 
information criterion (PSIS-LOO) for an object of class \code{lmBayes}. 
The PSIS-LOO is converted into deviance scale so that it is comparable 
with other information criteria like AIC and BIC.
}
\author{
Santiago Marin
}
