% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_class_functions.R
\name{residuals.lmBayes}
\alias{residuals.lmBayes}
\alias{resid.lmBayes}
\title{Residuals from an object of class \code{lmBayes}}
\usage{
\method{residuals}{lmBayes}(object, ...)
}
\arguments{
\item{object}{An object of class 'lmBayes'.}

\item{...}{Further arguments passed to.}
}
\value{
A data.frame with summary statistics of the posterior predictive 
distribution of the model residuals.
}
\description{
Summary statistics of the posterior predictive distribution of the 
residuals from an object of class \code{lmBayes}.
}
\author{
Santiago Marin
}
