% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{summary.lmBayes}
\alias{summary.lmBayes}
\title{Summary table of the results for an object of class \code{'lmBayes'}}
\usage{
\method{summary}{lmBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{'lmBayes'}.}

\item{...}{Further arguments passed to.}
}
\description{
Summary table of the results for an object of class \code{'lmBayes'}
}
\author{
Santiago Marin
}
