% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{summary.spmBayes}
\alias{summary.spmBayes}
\title{Summary table of the results for an object of class \code{spmBayes}}
\usage{
\method{summary}{spmBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{spmBayes}.}

\item{...}{Further arguments passed to.}
}
\description{
Summary table of the results for an object of class \code{spmBayes}
}
\author{
Santiago Marin
}
