% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{widely.aic}
\alias{widely.aic}
\title{WAIC}
\usage{
widely.aic(object)
}
\arguments{
\item{object}{An object of class 'lmBayes'.}
}
\value{
The estimated WAIC.
}
\description{
This function computes the Watanabe–Akaike information criterion (WAIC)
for an object of class \code{lmBayes}. The WAIC is converted into deviance 
scale so that it is comparable with other information criteria 
like AIC and BIC.
}
\author{
Santiago Marin
}
