% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnplasso-package.R
\docType{package}
\name{bnplasso}
\alias{bnplasso-package}
\alias{bnplasso}
\title{bnplasso: Fit linear regression models with the Nonparametric Bayesian Lasso}
\description{
Fit linear regression models with the Nonparametric Bayesian Lasso, a novel 
and adaptive shrinkage prior for Bayesian regression and variable selection. 
Posterior inference is conducted via Markov chain Monte Carlo (MCMC) as 
described in Marin et al. (2025+) 
<https://doi.org/10.1080/10618600.2025.2572327>.
}
\references{
S. Marin, B. Long,and A. H. Westveld (2025+), Adaptive Shrinkage with a 
Nonparametric Bayesian Lasso. \emph{Journal of Computational and Graphical 
Statistics}. doi:10.1080/10618600.2025.2572327
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marinsantiago/bnplasso}
  \item Report bugs at \url{https://github.com/marinsantiago/bnplasso/issues}
}

}
\author{
\strong{Maintainer}: Santiago Marin \email{santiago.marinardila@anu.edu.au} (\href{https://orcid.org/0009-0006-0086-3721}{ORCID})

Other contributors:
\itemize{
  \item Bronwyn Loong \email{bronwyn.loong@anu.edu.au} (\href{https://orcid.org/0000-0002-1409-8892}{ORCID}) [contributor]
  \item Anton H. Westveld \email{anton.westveld@anu.edu.au} (\href{https://orcid.org/0000-0003-1671-0931}{ORCID}) [contributor]
}

}
