% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_class_functions.R
\name{coef.lmBayes}
\alias{coef.lmBayes}
\alias{coefficients.lmBayes}
\title{Regression coefficients from an object of class \code{lmBayes}.}
\usage{
\method{coef}{lmBayes}(object, ...)
}
\arguments{
\item{object}{An object of class 'lmBayes'.}

\item{...}{Further arguments passed to.}
}
\value{
A data.frame with summary statistics of the posterior distribution of
the regression coefficients.
}
\description{
Summary statistics of the posterior distribution of the regression 
coefficients from an object of class \code{lmBayes}.
}
\author{
Santiago Marin
}
