% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_methods.R
\name{fitted.lmBayes}
\alias{fitted.lmBayes}
\alias{fitted.values.lmBayes}
\title{Fitted Values from an object of class \code{lmBayes}}
\usage{
\method{fitted}{lmBayes}(object, ...)
}
\arguments{
\item{object}{An object of class \code{'lmBayes'}.}

\item{...}{Further arguments passed to.}
}
\value{
A data.frame with summary statistics of the posterior predictive 
distribution of the fitted values.
}
\description{
Summary statistics of the posterior predictive distribution of the 
fitted values from an object of class \code{lmBayes}.
}
\author{
Santiago Marin
}
