% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3_class_functions.R
\name{print.lmBayes}
\alias{print.lmBayes}
\title{Print the results for an object of class \code{'lmBayes'}.}
\usage{
\method{print}{lmBayes}(x, ...)
}
\arguments{
\item{x}{An object of class 'lmBayes'.}

\item{...}{Further arguments passed to.}
}
\description{
Print the results for an object of class \code{'lmBayes'}.
}
\author{
Santiago Marin
}
